fix(pop): update Malta's population to official estimate
Malta's population estimate in the 2019 edition of the UNWPP dataset is much lower (441,539) than estimates published by other organizations such as Eurostat (516,100), and especially than the denominator used in the official GitHub repository of Malta's government (514,564), which is also used by the European CDC for its vaccination data. For this reason, we're updating the estimate of Malta's population to 514,564.