fix taiwan.py: support of new formats
as of 2021/11/29 the embedded table has changed from 4-column to be
3-column, with the last column containing two numbers separated by
whitespaces.

I'm not sure whether this change is going to be permanent or just a
glitch... instead of stop supporting old format, it's probably safer
to extend existing parser of _parse_stats to support both the new and
old format for a little while.