fix taiwan: pdf link extraction.
Usually the PDF URLs harvested from the download page[1] would respond
an HTML UI that is the viewer of the actual pdf with a download link,
but occassionally the same URL gives directly leads to the actual pdf,
without revealing the viewer.

Such behaviour seems to be decided on the server-side. Luckily enough
though, the server properly sents the correct Content-Type so we can
basiaclly assume that when it is "application/pdf", we found the pdf url.

The column names are also changed -- the "dates" now contain 3-digits
that is the ROC year.

[1]: https://www.cdc.gov.tw/Category/Page/9jFXNbCe-sFK9EImRRi2Og